+++ apps/api/views/user_views.py
@@ -1,12 +1,18 @@
 import logging
 import sys
 
+from logging import Logger
+
+from django.core.handlers.wsgi import WSGIRequest
+from django.db.models import QuerySet
+from django.http import QueryDict
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework_simplejwt.exceptions import InvalidToken, TokenError
 
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
+from rest_framework.serializers import Serializer
 from rest_framework import status
 
 from django.core.exceptions import ObjectDoesNotExist
@@ -16,17 +22,19 @@ from drf_yasg.utils import swagger_auto_schema
 
 from ..serializers import MyTokenObtainPairSerializer, UserSerializerWithToken, UserSerializer
 
+from ...core.models import User
+
 from ..services.user_services import create_user_by_data_service, get_user_by_slug_service, verify_user_service, \
     generate_new_verify_code_for_user_service, verify_user_by_code_service
 
 from apps.tasks.tasks import send_email_task
 
-logger = logging.getLogger('django')
+logger: Logger = logging.getLogger('django')
 
 
 class MyTokenObtainPairView(TokenObtainPairView):
     """Контроллер для авторизации"""
-    serializer_class = MyTokenObtainPairSerializer
+    serializer_class: Serializer[MyTokenObtainPairSerializer] = MyTokenObtainPairSerializer
 
     @swagger_auto_schema(
         manual_parameters=[
@@ -63,8 +71,8 @@ class MyTokenObtainPairView(TokenObtainPairView):
         operation_description='Данный endpoint позволяет авторизовать пользователя.',
         operation_summary='Авторизовать пользователя'
     )
-    def post(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
+    def post(self, request: WSGIRequest, *args, **kwargs) -> Response:
+        serializer: Serializer = self.get_serializer(data=request.data)
 
         try:
             serializer.is_valid(raise_exception=True)
@@ -128,17 +136,17 @@ class MyTokenObtainPairView(TokenObtainPairView):
     operation_summary='Зарегистрировать user'
 )
 @api_view(['POST'])
-def register_user_view(request):
+def register_user_view(request: WSGIRequest) -> Response:
     """Контроллер для регистрации новых пользователей"""
-    data = request.data
+    data: QueryDict = request.data
     try:
-        user = create_user_by_data_service(data)
-        serializer = UserSerializerWithToken(user, many=False)
-        message = 'Запрос выполнен успешно'
+        user: QuerySet[User] = create_user_by_data_service(data)
+        serializer: Serializer[UserSerializerWithToken] = UserSerializerWithToken(user, many=False)
+        message: str = 'Запрос выполнен успешно'
         logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
-        subject = 'Администратор Rectop'
-        body = f'Для активации аккаунта перейдите по ссылке http://127.0.0.1:3000/auth/verify/{user.slug} \n ' \
-               f'Спасибо за регистрацию!'
+        subject: str = 'Администратор Rectop'
+        body: str = f'Для активации аккаунта перейдите по ссылке http://127.0.0.1:3000/auth/verify/{user.slug} \n ' \
+                    f'Спасибо за регистрацию!'
         send_email_task.delay(subject, body, user.email.__str__())
         return Response(
             data=serializer.data,
@@ -146,7 +154,7 @@ def register_user_view(request):
         )
 
     except KeyError as e:
-        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        message: str = f'Ошибка при обработке запроса. Отсутствует поле {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -156,7 +164,7 @@ def register_user_view(request):
         )
 
     except Exception as e:
-        message = f'Пользователем с таким email уже существует {e}'
+        message: str = f'Пользователем с таким email уже существует {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -188,10 +196,10 @@ def register_user_view(request):
 )
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def read_user_profile_view(request):
+def read_user_profile_view(request: WSGIRequest) -> Response:
     """Контроллер для отдачи информации о текущем пользователе"""
-    user = request.user
-    serializer = UserSerializer(user, many=False)
+    user: QuerySet[User] = request.user
+    serializer: Serializer[UserSerializer] = UserSerializer(user, many=False)
     logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - Запрос выполнен успешно')
     return Response(
         data=serializer.data
@@ -274,15 +282,15 @@ def read_user_profile_view(request):
 )
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated])
-def update_user_profile_view(request):
+def update_user_profile_view(request: WSGIRequest) -> Response:
     """Контроллер обновления пользовательских настроек"""
-    user = request.user
-    data = request.data
+    user: QuerySet[User] = request.user
+    data: QueryDict = request.data
     try:
-        serializer = UserSerializer(user, many=False, partial=True, data=data)
+        serializer: Serializer[UserSerializer] = UserSerializer(user, many=False, partial=True, data=data)
         if serializer.is_valid():
             serializer.save()
-        message = 'Запрос выполнен успешно'
+        message: str = 'Запрос выполнен успешно'
         logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data=serializer.data,
@@ -290,7 +298,7 @@ def update_user_profile_view(request):
         )
 
     except KeyError as e:
-        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        message: str = f'Ошибка при обработке запроса. Отсутствует поле {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -300,7 +308,7 @@ def update_user_profile_view(request):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={
@@ -334,21 +342,21 @@ def update_user_profile_view(request):
     operation_summary='Верифицировать user'
 )
 @api_view(['GET'])
-def verify_user_view(request, slug):
+def verify_user_view(request: WSGIRequest, slug: str) -> Response:
     """Контроллер для верификации пользователя"""
     try:
-        user = get_user_by_slug_service(slug)
+        user: QuerySet[User] = get_user_by_slug_service(slug)
         logger.error(f'!!!!!!!!!!!! {user}')
-        user = verify_user_service(user)
-        serializer = UserSerializer(user, many=False)
-        message = 'Запрос выполнен успешно'
+        user: QuerySet[User] = verify_user_service(user)
+        serializer: Serializer[UserSerializer] = UserSerializer(user, many=False)
+        message: str = 'Запрос выполнен успешно'
         logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data=serializer.data,
             status=status.HTTP_200_OK
         )
     except ObjectDoesNotExist as e:
-        message = f'Такой slug не найден {e}'
+        message: str = f'Такой slug не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -358,7 +366,7 @@ def verify_user_view(request, slug):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -392,23 +400,23 @@ def verify_user_view(request, slug):
 )
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
-def send_verify_code_user_view(request):
+def send_verify_code_user_view(request: WSGIRequest) -> Response:
     """Контроллер для отправки кода верификации пользователя"""
     try:
-        user = request.user
+        user: QuerySet[User] = request.user
         user = generate_new_verify_code_for_user_service(user.id)
-        subject = 'Администратор Rectop'
-        body = f'Приветствуем вас, {user.get_full_name()}! \n Ваш код подтверждения: {user.verify_code}'
-        email = user.email.__str__()
+        subject: str = 'Администратор Rectop'
+        body: str = f'Приветствуем вас, {user.get_full_name()}! \n Ваш код подтверждения: {user.verify_code}'
+        email: str = user.email.__str__()
         send_email_task.delay(subject, body, email)
-        message = 'Запрос выполнен успешно'
+        message: str = 'Запрос выполнен успешно'
         logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(
             data={'detail': message},
             status=status.HTTP_200_OK
         )
     except ObjectDoesNotExist as e:
-        message = f'Такой user не найден {e}'
+        message: str = f'Такой user не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -418,7 +426,7 @@ def send_verify_code_user_view(request):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -467,20 +475,20 @@ def send_verify_code_user_view(request):
 )
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
-def verify_code_user_view(request):
+def verify_code_user_view(request: WSGIRequest) -> Response:
     """Контроллер для верификации пользователя по коду"""
     try:
-        user = request.user
-        verify_code = request.data['verify_code']
+        user: QuerySet[User] = request.user
+        verify_code: str = request.data['verify_code']
         if verify_user_by_code_service(user, verify_code):
-            message = 'Запрос выполнен успешно'
+            message: str = 'Запрос выполнен успешно'
             logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={'detail': message},
                 status=status.HTTP_200_OK
             )
         else:
-            message = 'Этот код не подходит'
+            message: str = 'Этот код не подходит'
             logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(
                 data={
@@ -489,7 +497,7 @@ def verify_code_user_view(request):
                 status=status.HTTP_403_FORBIDDEN
             )
     except ObjectDoesNotExist as e:
-        message = f'Такой user не найден {e}'
+        message: str = f'Такой user не найден {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -499,7 +507,7 @@ def verify_code_user_view(request):
         )
 
     except KeyError as e:
-        message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
+        message: str = f'Ошибка при обработке запроса. Отсутствует поле {e}'
         logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={
@@ -509,7 +517,7 @@ def verify_code_user_view(request):
         )
 
     except Exception as e:
-        message = f'Ошибка при обработке запроса {e}'
+        message: str = f'Ошибка при обработке запроса {e}'
         logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(
             data={